server/requirements.txt
flask==3.0
flask-cors
pymongo
bcrypt
pyjwt
